refactor: reorganize data-layer
