+++ apps/web/src/data/domains/stations/getStation.ts
@@ -1,4 +1,4 @@
-import db from "..";
+import db from "../../";
 
 const getStationById = async (stationId: string) => {
   db.query.stations.findFirst({
