--- apps/web/src/data/schema.ts
@@ -1,8 +1,0 @@
-export * from "./departments/schema";
-export * from "./stations/schema";
-export * from "./firetrucks/schema";
-export * from "./firefighters/schema";
-export * from "./beacons/schema";
-export * from "./qualifications/schema";
-export * from "./crew/schema";
-export * from "./incident/schema";
