+++ apps/web/src/data/shared/schema/index.ts
@@ -1,0 +1,8 @@
+export * from "./departments";
+export * from "./stations";
+export * from "./firetrucks";
+export * from "./firefighters";
+export * from "./beacons";
+export * from "./qualifications";
+export * from "./crews";
+export * from "./incidents";
